fix: change compateObgects to areObjectsEqual gf-89
