+++ apps/frontend/src/libs/helpers/helpers.ts
@@ -1,3 +1,3 @@
-export { compareObjects } from "./compare-objects/compare-objects.helper.js";
+export { areObjectsEqual } from "./are-objects-equal/are-objects-equal.helper.js";
 export { getValidClassNames } from "./get-valid-classes/get-valid-classes.helper.js";
 export { configureString, formatDate } from "@git-fit/shared";
